refactor(cli): split session command modules
