+++ src/info.rs
@@ -1,0 +1,190 @@
+use anyhow::Result;
+
+use crate::adapters;
+use crate::config::AppConfig;
+use crate::db::store::Store;
+
+struct SourceSummary {
+    label: String,
+    id: String,
+    sessions: usize,
+    messages: usize,
+    range: String,
+    error: Option<String>,
+}
+
+pub fn run() -> Result<()> {
+    let all = adapters::all_adapters();
+    let labels = adapters::source_labels();
+    let mut config = AppConfig::load_or_default();
+    config.normalize_sources(&labels);
+    let store = Store::open()?;
+    let progress = store.semantic_progress().unwrap_or_default();
+    let worker = store.background_job_status("pipeline").unwrap_or_default();
+
+    let mut rows = Vec::new();
+    let mut grand_sessions = 0usize;
+    let mut grand_messages = 0usize;
+
+    for adapter in &all {
+        let id = adapter.id();
+        let label =
+            labels.iter().find(|(k, _)| k == id).map(|(_, v)| v.as_str()).unwrap_or(id).to_string();
+
+        match adapter.scan_summary() {
+            Ok(Some(summary)) => {
+                grand_sessions += summary.sessions;
+                grand_messages += summary.messages;
+
+                rows.push(SourceSummary {
+                    label,
+                    id: id.to_string(),
+                    sessions: summary.sessions,
+                    messages: summary.messages,
+                    range: format_date_range(summary.oldest_started_at, summary.newest_started_at),
+                    error: None,
+                });
+            }
+            Ok(None) => match adapter.scan() {
+                Ok(sessions) => {
+                    let session_count = sessions.len();
+                    let message_count: usize = sessions.iter().map(|s| s.messages.len()).sum();
+                    let oldest = sessions.iter().map(|s| s.started_at).min();
+                    let newest = sessions.iter().map(|s| s.started_at).max();
+
+                    grand_sessions += session_count;
+                    grand_messages += message_count;
+
+                    rows.push(SourceSummary {
+                        label,
+                        id: id.to_string(),
+                        sessions: session_count,
+                        messages: message_count,
+                        range: format_date_range(oldest, newest),
+                        error: None,
+                    });
+                }
+                Err(e) => {
+                    rows.push(SourceSummary {
+                        label,
+                        id: id.to_string(),
+                        sessions: 0,
+                        messages: 0,
+                        range: "-".to_string(),
+                        error: Some(e.to_string()),
+                    });
+                }
+            },
+            Err(e) => {
+                rows.push(SourceSummary {
+                    label,
+                    id: id.to_string(),
+                    sessions: 0,
+                    messages: 0,
+                    range: "-".to_string(),
+                    error: Some(e.to_string()),
+                });
+            }
+        }
+    }
+
+    let source_width = rows
+        .iter()
+        .map(|row| format!("{} ({})", row.label, row.id).len())
+        .max()
+        .unwrap_or(12)
+        .max("Source".len());
+    let sessions_width = rows
+        .iter()
+        .map(|row| row.sessions.to_string().len())
+        .max()
+        .unwrap_or(1)
+        .max("Sessions".len())
+        .max(grand_sessions.to_string().len());
+    let messages_width = rows
+        .iter()
+        .map(|row| row.messages.to_string().len())
+        .max()
+        .unwrap_or(1)
+        .max("Messages".len())
+        .max(grand_messages.to_string().len());
+
+    println!("Source Scan");
+    println!(
+        "  {source:<source_width$}  {sessions:>sessions_width$}  {messages:>messages_width$}  Range",
+        source = "Source",
+        sessions = "Sessions",
+        messages = "Messages"
+    );
+    for row in rows {
+        let source = format!("{} ({})", row.label, row.id);
+        if let Some(error) = row.error {
+            println!(
+                "  {source:<source_width$}  {sessions:>sessions_width$}  {messages:>messages_width$}  error: {error}",
+                sessions = "-",
+                messages = "-"
+            );
+            continue;
+        }
+        println!(
+            "  {source:<source_width$}  {sessions:>sessions_width$}  {messages:>messages_width$}  {range}",
+            sessions = row.sessions,
+            messages = row.messages,
+            range = row.range
+        );
+    }
+    println!(
+        "  {source:<source_width$}  {sessions:>sessions_width$}  {messages:>messages_width$}",
+        source = "Total scanned",
+        sessions = grand_sessions,
+        messages = grand_messages
+    );
+
+    println!();
+    println!("Settings");
+    println!(
+        "  Sources     {}",
+        labels
+            .iter()
+            .filter(|(id, _)| config.is_source_enabled(id))
+            .map(|(_, label)| label.as_str())
+            .collect::<Vec<_>>()
+            .join(", ")
+    );
+    println!("  Time scope  {}", config.sync_window.label());
+
+    println!();
+    println!("Semantic Queue");
+    println!("  Indexed DB  {} sessions tracked locally", progress.total_sessions);
+    println!(
+        "  Progress    {} done, {} pending, {} failed",
+        progress.done_sessions,
+        progress.pending_sessions + progress.processing_sessions,
+        progress.failed_sessions
+    );
+    if let Some(phase) = worker.phase {
+        println!("  Worker      {phase}");
+    }
+
+    println!();
+    println!("Tip: open the TUI and press Ctrl+S to edit settings.");
+
+    Ok(())
+}
+
+fn format_date_range(oldest: Option<i64>, newest: Option<i64>) -> String {
+    if oldest.is_none() && newest.is_none() {
+        return "-".to_string();
+    }
+
+    let oldest = oldest
+        .and_then(chrono::DateTime::from_timestamp_millis)
+        .map(|dt| dt.format("%Y-%m-%d").to_string())
+        .unwrap_or_else(|| "-".to_string());
+    let newest = newest
+        .and_then(chrono::DateTime::from_timestamp_millis)
+        .map(|dt| dt.format("%Y-%m-%d").to_string())
+        .unwrap_or_else(|| "-".to_string());
+
+    format!("{oldest} -> {newest}")
+}
