+++ src/lib.rs
@@ -5,9 +5,15 @@ pub mod db;
 pub mod embedding;
 pub mod export;
 pub mod import;
+pub mod info;
+pub mod query;
 pub mod semantic;
+pub mod session_action;
 pub mod share;
+pub mod share_init;
 pub mod skill_audit;
+pub mod sync;
+pub mod transcript;
 pub mod tui;
 pub mod types;
 pub mod usage;
