+++ src/tui/mod.rs
@@ -1,3 +1,4 @@
 pub mod app;
 pub mod event;
+pub mod runner;
 pub mod ui;
